refactor: centralize decision constants in const
