+++ src/context_compiler/__init__.py
@@ -1,8 +1,13 @@
 from importlib.metadata import version
 
-from .const import POLICY_PROHIBIT, POLICY_USE
+from .const import (
+    DECISION_CLARIFY,
+    DECISION_PASSTHROUGH,
+    DECISION_UPDATE,
+    POLICY_PROHIBIT,
+    POLICY_USE,
+)
 from .controller import PreviewResult, StepResult, StructuralDiff, preview, state_diff, step
-from .decision_constants import DECISION_CLARIFY, DECISION_PASSTHROUGH, DECISION_UPDATE
 from .decision_helpers import (
     get_clarify_prompt,
     get_decision_state,
