+++ src/context_compiler/decision_constants.py
@@ -1,5 +1,9 @@
 """Public decision-kind string constants for host-side branching ergonomics."""
 
-DECISION_PASSTHROUGH = "passthrough"
-DECISION_UPDATE = "update"
-DECISION_CLARIFY = "clarify"
+from .const import DECISION_CLARIFY, DECISION_PASSTHROUGH, DECISION_UPDATE
+
+__all__ = [
+    "DECISION_PASSTHROUGH",
+    "DECISION_UPDATE",
+    "DECISION_CLARIFY",
+]
