+++ src/context_compiler/repl.py
@@ -5,10 +5,10 @@ from typing import TextIO
 from experimental.preprocessor.output_validation import parse_preprocessor_output
 
 from . import __version__, create_engine, get_policy_items, get_premise_value
+from .const import DECISION_CLARIFY, DECISION_PASSTHROUGH
 from .controller import OUTPUT_VERSION, PreviewResult, StepResult
 from .controller import preview as controller_preview
 from .controller import step as controller_step
-from .decision_constants import DECISION_CLARIFY, DECISION_PASSTHROUGH
 from .engine import Decision, DecisionKind, Engine, State
 
 _EXIT_TOKENS = {"exit", "quit"}
